test: Move radix tree stress test from unit to boost
Just move the code. Test "scale" is also taken from default unit test
arguments.
