+++ test/unit/CMakeLists.txt
@@ -8,8 +8,6 @@ add_scylla_test(lsa_sync_eviction_test
   KIND UNIT)
 add_scylla_test(lsa_async_eviction_test
   KIND UNIT)
-add_scylla_test(radix_tree_stress_test
-  KIND UNIT)
 add_scylla_test(radix_tree_compaction_test
   KIND UNIT)
 add_scylla_test(row_cache_alloc_stress_test
